NO-EVE<CRIAÇÃO DE EVENTO>

+++ src/main/java/com/alphasolutions/eventapi/repository/EventoRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface EventoRepository extends JpaRepository<Evento, Integer> {
+public interface EventoRepository extends JpaRepository<Evento, Long> {
 
     Evento findByIdEvento(Long idEvento);
 }
